AbstractAllocateDecision and much code reduction as a result
